+++ server/src/lib.rs
@@ -35,6 +35,7 @@ extern crate alloc;
 mod future;
 pub mod middleware;
 mod server;
+mod transport;
 
 #[cfg(test)]
 mod tests;
