+++ server/src/transport/http.rs
@@ -1,32 +1,31 @@
-pub mod response;
-
+use std::convert::Infallible;
 use std::net::SocketAddr;
 
-use crate::error::GenericTransportError;
-use crate::http_helpers::{self, read_body};
-use crate::logger::{self, HttpLogger as Logger};
-use crate::server::helpers::{prepare_error, BatchResponse, BatchResponseBuilder, MethodResponse};
-use crate::server::rpc_module::MethodKind;
-use crate::server::{resource_limiting::Resources, rpc_module::Methods};
-use crate::tracing::{rx_log_from_json, tx_log_from_str, RpcTracing};
-use crate::JsonRawValue;
 use futures_util::TryStreamExt;
 use http::Method;
+use jsonrpsee_core::error::GenericTransportError;
+use jsonrpsee_core::http_helpers::{self, read_body};
+use jsonrpsee_core::logger::{self, HttpLogger as Logger};
+use jsonrpsee_core::server::helpers::{prepare_error, BatchResponse, BatchResponseBuilder, MethodResponse};
+use jsonrpsee_core::server::rpc_module::MethodKind;
+use jsonrpsee_core::server::{resource_limiting::Resources, rpc_module::Methods};
+use jsonrpsee_core::tracing::{rx_log_from_json, tx_log_from_str, RpcTracing};
+use jsonrpsee_core::JsonRawValue;
 use jsonrpsee_types::error::{ErrorCode, BATCHES_NOT_SUPPORTED_CODE, BATCHES_NOT_SUPPORTED_MSG};
 use jsonrpsee_types::{ErrorObject, Id, Notification, Params, Request};
 use tracing_futures::Instrument;
 
-use super::access_control::AccessControl;
+use jsonrpsee_core::server::access_control::AccessControl;
 
 type Notif<'a> = Notification<'a, Option<&'a JsonRawValue>>;
 
 /// Checks that content type of received request is valid for JSON-RPC.
-pub fn content_type_is_json(request: &hyper::Request<hyper::Body>) -> bool {
+pub(crate) fn content_type_is_json(request: &hyper::Request<hyper::Body>) -> bool {
 	is_json(request.headers().get("content-type"))
 }
 
 /// Returns true if the `content_type` header indicates a valid JSON message.
-pub fn is_json(content_type: Option<&hyper::header::HeaderValue>) -> bool {
+pub(crate) fn is_json(content_type: Option<&hyper::header::HeaderValue>) -> bool {
 	match content_type.and_then(|val| val.to_str().ok()) {
 		Some(content)
 			if content.eq_ignore_ascii_case("application/json")
@@ -39,21 +38,34 @@ pub fn is_json(content_type: Option<&hyper::header::HeaderValue>) -> bool {
 	}
 }
 
+pub(crate) async fn reject_connection(socket: tokio::net::TcpStream) {
+	async fn reject(_req: hyper::Request<hyper::Body>) -> Result<hyper::Response<hyper::Body>, Infallible> {
+		Ok(response::too_many_requests())
+	}
+
+	if let Err(e) = hyper::server::conn::Http::new().serve_connection(socket, hyper::service::service_fn(reject)).await
+	{
+		tracing::warn!("Error when trying to deny connection: {:?}", e);
+	}
+}
+
 #[derive(Debug)]
-pub struct ProcessValidatedRequest<L: Logger> {
-	pub request: hyper::Request<hyper::Body>,
-	pub logger: L,
-	pub methods: Methods,
-	pub resources: Resources,
-	pub max_request_body_size: u32,
-	pub max_response_body_size: u32,
-	pub max_log_length: u32,
-	pub batch_requests_supported: bool,
-	pub request_start: L::Instant,
+pub(crate) struct ProcessValidatedRequest<L: Logger> {
+	pub(crate) request: hyper::Request<hyper::Body>,
+	pub(crate) logger: L,
+	pub(crate) methods: Methods,
+	pub(crate) resources: Resources,
+	pub(crate) max_request_body_size: u32,
+	pub(crate) max_response_body_size: u32,
+	pub(crate) max_log_length: u32,
+	pub(crate) batch_requests_supported: bool,
+	pub(crate) request_start: L::Instant,
 }
 
 /// Process a verified request, it implies a POST request with content type JSON.
-pub async fn process_validated_request<L: Logger>(input: ProcessValidatedRequest<L>) -> hyper::Response<hyper::Body> {
+pub(crate) async fn process_validated_request<L: Logger>(
+	input: ProcessValidatedRequest<L>,
+) -> hyper::Response<hyper::Body> {
 	let ProcessValidatedRequest {
 		request,
 		logger,
@@ -123,13 +135,13 @@ pub async fn process_validated_request<L: Logger>(input: ProcessValidatedRequest
 }
 
 #[derive(Debug, Clone)]
-pub struct Batch<'a, L: Logger> {
+pub(crate) struct Batch<'a, L: Logger> {
 	data: Vec<u8>,
 	call: CallData<'a, L>,
 }
 
 #[derive(Debug, Clone)]
-pub struct CallData<'a, L: Logger> {
+pub(crate) struct CallData<'a, L: Logger> {
 	conn_id: usize,
 	logger: &'a L,
 	methods: &'a Methods,
@@ -140,7 +152,7 @@ pub struct CallData<'a, L: Logger> {
 }
 
 #[derive(Debug, Clone)]
-pub struct Call<'a, L: Logger> {
+pub(crate) struct Call<'a, L: Logger> {
 	params: Params<'a>,
 	name: &'a str,
 	call: CallData<'a, L>,
@@ -150,7 +162,7 @@ pub struct Call<'a, L: Logger> {
 // Batch responses must be sent back as a single message so we read the results from each
 // request in the batch and read the results off of a new channel, `rx_batch`, and then send the
 // complete batch response back to the client over `tx`.
-pub async fn process_batch_request<L>(b: Batch<'_, L>) -> BatchResponse
+pub(crate) async fn process_batch_request<L>(b: Batch<'_, L>) -> BatchResponse
 where
 	L: Logger,
 {
@@ -199,7 +211,7 @@ where
 	BatchResponse::error(id, ErrorObject::from(code))
 }
 
-pub async fn process_single_request<L: Logger>(data: Vec<u8>, call: CallData<'_, L>) -> MethodResponse {
+pub(crate) async fn process_single_request<L: Logger>(data: Vec<u8>, call: CallData<'_, L>) -> MethodResponse {
 	if let Ok(req) = serde_json::from_slice::<Request>(&data) {
 		let trace = RpcTracing::method_call(&req.method);
 		async {
@@ -224,7 +236,7 @@ pub async fn process_single_request<L: Logger>(data: Vec<u8>, call: CallData<'_,
 	}
 }
 
-pub async fn execute_call<L: Logger>(c: Call<'_, L>) -> MethodResponse {
+pub(crate) async fn execute_call<L: Logger>(c: Call<'_, L>) -> MethodResponse {
 	let Call { name, id, params, call } = c;
 	let CallData { resources, methods, logger, max_response_body_size, max_log_length, conn_id, request_start } = call;
 
@@ -277,19 +289,19 @@ pub async fn execute_call<L: Logger>(c: Call<'_, L>) -> MethodResponse {
 	response
 }
 
-pub struct HandleRequest<L: Logger> {
-	pub remote_addr: SocketAddr,
-	pub methods: Methods,
-	pub acl: AccessControl,
-	pub resources: Resources,
-	pub max_request_body_size: u32,
-	pub max_response_body_size: u32,
-	pub max_log_length: u32,
-	pub batch_requests_supported: bool,
-	pub logger: L,
+pub(crate) struct HandleRequest<L: Logger> {
+	pub(crate) remote_addr: SocketAddr,
+	pub(crate) methods: Methods,
+	pub(crate) acl: AccessControl,
+	pub(crate) resources: Resources,
+	pub(crate) max_request_body_size: u32,
+	pub(crate) max_response_body_size: u32,
+	pub(crate) max_log_length: u32,
+	pub(crate) batch_requests_supported: bool,
+	pub(crate) logger: L,
 }
 
-pub async fn handle_request<L: Logger>(
+pub(crate) async fn handle_request<L: Logger>(
 	request: hyper::Request<hyper::Body>,
 	input: HandleRequest<L>,
 ) -> hyper::Response<hyper::Body> {
@@ -342,5 +354,99 @@ pub async fn handle_request<L: Logger>(
 		// Error scenarios:
 		Method::POST => response::unsupported_content_type(),
 		_ => response::method_not_allowed(),
+	}
+}
+
+pub(crate) mod response {
+	use jsonrpsee_types::error::reject_too_big_request;
+	use jsonrpsee_types::error::{ErrorCode, ErrorResponse};
+	use jsonrpsee_types::Id;
+
+	const JSON: &str = "application/json; charset=utf-8";
+	const TEXT: &str = "text/plain";
+
+	/// Create a response for json internal error.
+	pub(crate) fn internal_error() -> hyper::Response<hyper::Body> {
+		let error = serde_json::to_string(&ErrorResponse::borrowed(ErrorCode::InternalError.into(), Id::Null))
+			.expect("built from known-good data; qed");
+
+		from_template(hyper::StatusCode::INTERNAL_SERVER_ERROR, error, JSON)
+	}
+
+	/// Create a text/plain response for not allowed hosts.
+	pub(crate) fn host_not_allowed() -> hyper::Response<hyper::Body> {
+		from_template(hyper::StatusCode::FORBIDDEN, "Provided Host header is not whitelisted.\n".to_owned(), TEXT)
+	}
+
+	/// Create a text/plain response for disallowed method used.
+	pub(crate) fn method_not_allowed() -> hyper::Response<hyper::Body> {
+		from_template(
+			hyper::StatusCode::METHOD_NOT_ALLOWED,
+			"Used HTTP Method is not allowed. POST or OPTIONS is required\n".to_owned(),
+			TEXT,
+		)
+	}
+
+	/// Create a text/plain response for rejected "Origin" headers.
+	pub(crate) fn origin_rejected(origin: Option<&str>) -> hyper::Response<hyper::Body> {
+		from_template(
+			hyper::StatusCode::FORBIDDEN,
+			format!("Origin: `{}` is not whitelisted.\n", origin.unwrap_or("")),
+			TEXT,
+		)
+	}
+
+	/// Create a json response for oversized requests (413)
+	pub(crate) fn too_large(limit: u32) -> hyper::Response<hyper::Body> {
+		let error = serde_json::to_string(&ErrorResponse::borrowed(reject_too_big_request(limit), Id::Null))
+			.expect("built from known-good data; qed");
+
+		from_template(hyper::StatusCode::PAYLOAD_TOO_LARGE, error, JSON)
+	}
+
+	/// Create a json response for empty or malformed requests (400)
+	pub(crate) fn malformed() -> hyper::Response<hyper::Body> {
+		let error = serde_json::to_string(&ErrorResponse::borrowed(ErrorCode::ParseError.into(), Id::Null))
+			.expect("built from known-good data; qed");
+
+		from_template(hyper::StatusCode::BAD_REQUEST, error, JSON)
+	}
+
+	/// Create a response body.
+	fn from_template<S: Into<hyper::Body>>(
+		status: hyper::StatusCode,
+		body: S,
+		content_type: &'static str,
+	) -> hyper::Response<hyper::Body> {
+		hyper::Response::builder()
+			.status(status)
+			.header("content-type", hyper::header::HeaderValue::from_static(content_type))
+			.body(body.into())
+			// Parsing `StatusCode` and `HeaderValue` is infalliable but
+			// parsing body content is not.
+			.expect("Unable to parse response body for type conversion")
+	}
+
+	/// Create a valid JSON response.
+	pub(crate) fn ok_response(body: String) -> hyper::Response<hyper::Body> {
+		from_template(hyper::StatusCode::OK, body, JSON)
+	}
+
+	/// Create a response for unsupported content type.
+	pub(crate) fn unsupported_content_type() -> hyper::Response<hyper::Body> {
+		from_template(
+			hyper::StatusCode::UNSUPPORTED_MEDIA_TYPE,
+			"Supplied content type is not allowed. Content-Type: application/json is required\n".to_owned(),
+			TEXT,
+		)
+	}
+
+	/// Create a response for when the server is busy and can't accept more requests.
+	pub(crate) fn too_many_requests() -> hyper::Response<hyper::Body> {
+		from_template(
+			hyper::StatusCode::TOO_MANY_REQUESTS,
+			"Too many connections. Please try again later.".to_owned(),
+			TEXT,
+		)
 	}
 }
